[refactor] 속성 생성 팩토리들 추상화
- 팩토리 수가 많아서 추상화 코드 작성
