+++ DrawingApp/DrawingApp/Factory/RectangleFactory.swift
@@ -1,24 +1,19 @@
 import Foundation
 
 class RectangleFactory: RandomFigureFactory {
-    let idFactory: RandomIdFactory
-    let colorFactory: RandomColorFactory
-    let pointFactory: RandomPointFactory
-    let alphaFactory: RandomAlphaFactory
+    private let propertyFactory: RandomValueFactory
     
     init(deviceSafeArea: Point, deviceScreenSize: Size) {
-        idFactory = RandomIdFactory()
-        colorFactory = RandomColorFactory()
-        pointFactory = RandomPointFactory(safeArea: deviceSafeArea, screen: deviceScreenSize)
-        alphaFactory = RandomAlphaFactory()
+        propertyFactory = RandomValueFactory( safeArea: deviceSafeArea,
+                                              screen: deviceScreenSize )
     }
     
     func create() -> RandomFigure {
         let size = Size(width: 150, height: 120)
-        let id = idFactory.create()
-        let point = pointFactory.create()
-        let color = colorFactory.create()
-        let alpha = alphaFactory.create()
+        let id = propertyFactory.createId()
+        let point = propertyFactory.createPoint()
+        let color = propertyFactory.createColor()
+        let alpha = propertyFactory.createAlpha()
         return Rectangle(size: size, id: id, point: point, color: color, alpha: alpha)
     }
 }
