--- DrawingApp/DrawingApp/RandomFigure.swift
@@ -1,4 +1,0 @@
-protocol RandomFigure: CustomStringConvertible {
-    var color: Color { get }
-    var alpha: Alpha { get }
-}
