--- DrawingApp/DrawingApp/RandomFigureFactory.swift
@@ -1,3 +1,0 @@
-protocol RandomFigureFactory {
-    func create() -> RandomFigure
-}
