--- DrawingApp/DrawingApp/Size.swift
@@ -1,15 +1,0 @@
-struct Size {
-    let width: Double
-    let height: Double
-    
-    init(width: Double, height: Double) {
-        self.width = width
-        self.height = height
-    }
-}
-
-extension Size: CustomStringConvertible {
-    var description: String {
-        return "W\(width), H\(height)"
-    }
-}
